+++ src/Stride.Shaders.Parsing/Stride.Shaders.Parsing.csproj
@@ -5,6 +5,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <LangVersion>latest</LangVersion>
+    
   </PropertyGroup>
 
   <ItemGroup>
